[0435] Added edit functions for apprenticeship
